Reorganized StapGraph's setLevelLimits -- setLevelLimits(id) now calls setTopLevelTo(level).
Added some stuff to MouseWheelListener to add/remove nodes in real-time when the user uses the mouse-wheel in Box view the graph will automatically add nodes to the top and bottom. For Tree view, if the user scrolls up we will automatically start following the parent of the RootNode, but nothing happens when we scroll down. I'd like to make it so that if you scroll down in Tree view it follows all children of the current root node... This will require some serious modifications to Tree view, so it'll have to wait for another commit.