Implemented scrolling down for tree view.
Caveat:

Scrolling down will not change the root node. This is because for any node with N children I have no idea which node the user wants to focus on, and arbitrarily assigning a root node is dangerous behaviour.

On a more serious 'please fix' note, scrolling down also just adds to the bottom without removing from the top.