Added error check for debuginfo to errors.prop. Added comments to error for stapdev to include instructions for debuginfo and make uprobes. (This should be the first error that a user sees, so we may as well give them all the solutions at once.
TODO: Get rid of the popup. It's really annoying. Just append the information to the console instead.

Modified SystemTapGraphTest to test save and open as well. Added abstract saveData function to SystemTapParser, implemented in StapGraphParser (this is for testing purposes, and also because it sort of makes sense to have it there).

Added checks for negative indices in SystemTalErrorHandler.