Revamped SystemTapParser's realtime parsing capability:
If realtime mode is active, the run() method will create and schedule a RealTimeJob. The run method of the RealTimeJob consists of repeated calls to the abstract realTimeParsing() function. All an implementing class needs to do is create an implementation for realTimeParsing() and ensure that the isDone flag is set to true when the process terminates. <CUrrently working on removing the latter restriction>

Roland: I removed the function you wrote for now, it's in my gedit -- when I get access to my sleepingthreads test plugin I'll try putting this into the SleepingThreadsParser class.