Add ability to create the proxy using a URI in addition to an IProject
In order to implement the RemoteConnection class in terms of this proxy,
we need to have the ability to create a proxy based upon the URI of
an executable, which may reside on a remote machine, but not the machine
where the project is located.  A similar situation occurs when downloading
a locally-built executable to a remote machine to run it.  In these cases
we need to access a remote file system which is not related to where the
project is located.

In addition to these changes, I also sprinkled in a few @Override annotations
on the methods of classes that implement an interface.  Note that I was
not able to do this on the methods in org.eclipse.linuxtools.rdt.proxy
because it uses Java 1.5 instead of 1.6, where @Override is not allowed
on methods which directly implement an interface (i.e. not from a superclass).

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>