Reimplement the RemoteConnection class in terms of the RemoteProxy
This version of RemoteConnection is a re-write and re-envision of the
RemoteConnection package that's in
org.eclipse.linuxtools.profiling.launch.remote plugin.  I intend for this
class to be the only mechanism used for connecting with the target
machine (whether it's local or remote), so there is no longer a need for a
separate .remote plugin.

The RemoteConnection class and related classes have been moved from
the .remote plugin to org.eclipse.linuxtools.profiling.launch, and are
now based upon IRemoteProxyManager, IRemoteCommandLauncher, and
IRemoteFileProxy.

The RemoteConnection class is basically now just a convenience class on
top of the above facilities, and is intended to make the job of writing
launchers a little easier.  We can add more methods over time as we find
the need.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>