Systemtap: Canceling process when Cancel button in pressed.
The cancel button in SelectServerDialog was not working for some tasks,
like running a systemtap script in IDE perspective.
This patch also improves the way this dialog deals with cancel. It was
not storing this in a preferences anymore, it is now returning a boolean
value to show if it was really cancelled.