tmf: Enforce existing sv types when reusing attributes
Once we assign an Integer or String statevalue to an attribute, make
sure every subsequent modification keeps using statevalues of the
same type.

This is a new restriction that was not there before, but it makes
sense to add it to the design. It also helps to isolate problems
in the state provider.

Fixes Eclipse bug #375878

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>