tmf: Ensure we don't return null on GSS queries
This would indicate a problem with the backend anyway, since
by design there should be no "holes" in a state history.

But since this can still happen, if there's a problem with the
backend or if the file is corrupted for example, then we should
catch (and log) the error here, and guarantee what is sent up
is non-null.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>