Add preselection of launch provider backend to launch provider's config tab.
When the launch provider's config tab is opened, a previously selected
provider is preselected; if no provider is found in the launch configuration,
the provider with the highest priority is preselected. This change required
refactoring that allowed for a fix of the launch configuration persistence
problem with the launch provider config tab to be made.

Change-Id: Ibca0dbc89a25563f130d4a33d17677f29edf8fc0
Reviewed-on: https://git.eclipse.org/r/7038
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>