Temporary fix API version for Kepler
The version in the master branch is correctly set to '1.2', the upcoming
version for that branch.

Since there are major API changes in the lttng-kepler branch and that
there is no '1.2' baseline yet, the API tool insists on a '2.0' version
being provided for these files.

The issue will be resolved LT 1.2 is released and we merge the master
branch in lttng-kepler: at that point everything should fall in place
as we have to undo these fixes to get rid of the API errors.


Change-Id: I3164f7d4facbd909540abb3dca66b3ba7627145a
Signed-off-by: Francois Chouinard <fchouinard@gmail.com>
Reviewed-on: https://git.eclipse.org/r/7174
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>