Fix BZ #389778 by tracking initialization state of all tabs.
The underlying LaunchConfigurationTabGroupViewer is internal and does
not have the API to add/remove tabs after being initialized. It is
unaware of any tabs added/removed and will call performApply() even if
they are still initializing. If performApply() is called during
initialization the configuration will be corrupted. It is necessary to
track the state of initialization for each tab to determine whether
performApply() can be safely called.

Change-Id: I5946276ed89ba1bb545e41667206ddad34519352
Reviewed-on: https://git.eclipse.org/r/7832
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
IP-Clean: Jeff Johnston <jjohnstn@redhat.com>
Tested-by: Jeff Johnston <jjohnstn@redhat.com>