Move handling of profile type plug-ins to the provider plug-in.
This commit alters how profile type provider plug-ins are handled. The
plug-in.xml files for each profiling type plug-in were kept. This was
done in order to keep the separate launch configurations, launchers and
preference pages as before. However, all functionality is now handled by
the Provider* classes, some of which implement the IExecutableExtension
interface, which is used to acquire attributes specified in a particular
extension point configuration. Because of this change the profiling type
plug-ins are no longer needed.

Change-Id: I37dfcaa70e8af5f22dfd6120d5a5c3293944d5e3
Reviewed-on: https://git.eclipse.org/r/7948
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>