tmf: Introduce the ITmfStatistics interface
With this, any type of ITmfTrace can now define its own
statistics provider. This can then be used to populate views
like the Histogram or the Statistics view.

A base implementation is proposed in TmfStatistics (and used by
TmfTrace), and it uses a state system for its backend. That
state system is completely contained in the TmfStatistics object,
so that it's not confused with any "real" state system that could
be present in the trace.

Change-Id: Ia0bf807cdbed28f9a6ebd47509412d4ea35f99cd
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7986