Check that user belongs to stapdev for callgraph views
Callgraph would throw NPE's if the user wasn't properly assigned to the
stapdev group.  Now we properly check that the user belongs to the
stapdev group, and if not, avoid running the views causing the NPE,
instead throwing an error message informing the user, and providing the
command to properly add the user to the correct group ($usermod -aG
stapdev user).  Strings have been properly externalized.

Change-Id: I36448b14207a157e69b85a866f9f613ad55fc8e2
Reviewed-on: https://git.eclipse.org/r/8100
Tested-by: Hudson CI
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>