Add functionality and API to run script locally
Running a systemtap script should not require the sshd service,
especially when simply trying to run the script locally.  This patch
adds the functionality to bypass the scp/ssh portions of the code and
simply run the command locally.  A get/set style api has been added to
modify and read a protected boolean representing the local/remote
choice.  The default remains to use sshd in order to keep the current
implemenation the same.  A section in the systemtap preferences should
be set up to modify and set the local/remote settings.

Change-Id: I93b69ff0e398bdc83c5fd7f7d280b0847fccd42a
Reviewed-on: https://git.eclipse.org/r/8164
Tested-by: Hudson CI
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>