tmf: Move the stats tree structure logic to the nodes
Re-organize the TmfStatisticsTree so that the information
about node children are stored in the nodes themselves,
and not all at the top-level. This makes debugging, and
eventually extending the tree, much easier, imo.

This patch also drops the fKeys (which limited the types
of children each node can have) and the weird limitation
where only nodes whose path depths are mutiples of 2 could
have values. Now the tree is more generic, to give better
flexibility to potential sub-classes.

Change-Id: I30bc0f7f47256388b6c55e09f09b26286723c0b9
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/8298
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>