tmf: Store the generic statistics timestamps in nanoseconds
Since the TmfStatistics can be used with any trace type, and not
just LTTng kernel traces, we can't know exactly what scale its
timestamps are in.

To be safe, convert all timestamps to nanoseconds before saving
them in the history database. When we want to display the
information, the query timestamps will also be converted to
nanoseconds.

Change-Id: Ibcd2fb309f01cdfeab6b480e41f740296694e4be
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/8242
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>