Fix time range synchronization issues
- ResourcesView is unnecessarily refreshing with initial window offset
on setFocus(). This is removed.

- TimeGraphViewer method resetStartFinishTime() causes time range
notification callback. The method is changed to call
setStartFinishTime() instead of setStartFinishTimeNotify().

- TimeGraphScale handling of mouse double-click is calling
resetStartFinishTime() expecting the notification. The notification is
now triggered explicitly by calling notifyStartFinishTime().

- HistogramView initial window span can be clipped if it is set early
when not enough events have been indexed in the trace. The full trace
histogram time range is now re-set on reception of TmfTraceUpdated
signal.

- The full trace histogram uses the zoom's duration in updateTimeRange()
on mouse up, and this duration could have been clipped. The method now
used the stored range duration, which does not get clipped.

Change-Id: Id480340933494855330fc6d4fae9429155efa88f
Reviewed-on: https://git.eclipse.org/r/8745
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>