Improve performance of event cache when scrolling up
When an event index is requested that is outside of the cache, the cache
population request always starts at the requested index and the cache
contains the following 'cache size' events. Therefore when scrolling up
outside of the cache boundaries, a new cache population request is
triggered at every scroll event.

This patch changes the behavior of the cache to increase the cache to
two blocks of 'cache size', and ajust the cache population request so
that the cache is centered on the requested index.

Change-Id: I890324a2f6d8afec5956ae73d66683eefc98d828
Reviewed-on: https://git.eclipse.org/r/9601
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>