Make the annotation column saner.
Title is removed as it was not really adding any value but repeating the
file name and adding coverage to the coverage column. This allows
simplifying the editor significantly as no layouting needs to be done.
Additionaly the way number of digits was calculated was awkward - it was
checking the length of the title and the number of lines. As the number
of lines in the document has no relation to the stuff in the annotation
it is changed to get the length of each annotation string and use the
biggest one.

Change-Id: I844db4d74cb023afa89265fb59b09bd80d63e5d0
Reviewed-on: https://git.eclipse.org/r/9928
Tested-by: Hudson CI
Reviewed-by: Xavier Raynaud <xavier.raynaud@kalray.eu>
IP-Clean: Xavier Raynaud <xavier.raynaud@kalray.eu>
Tested-by: Xavier Raynaud <xavier.raynaud@kalray.eu>