Use transparent color for unknown states in ControlFlow/Resources view
The TRANSPARENT value is used for unknown states (states embedded by the
view in between actual states due to the sampling). When many states
occur in the same pixel range, the pixel now takes the color of one of
the states within the pixel range. The selected state color for a pixel
can change with regards to the current widget size and time range. The
tool tip information may be computed for a transparent state which does
not match the visible color under the cursor. In that case the state
name "(multiple)" is displayed.

Change-Id: If9185f67826bdab450cecf6f70c8b25f9beda33d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/10187
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Hudson CI