lttng ui: fix handling of command output for verbose commands
When executing commands with preference verbose the command output
parsing of certain commands were not correct and caused
ExecutionExceptions in Eclipse even if the command was successfully
executed. This has been corrected with this commit. Also, this
commit contains JUnit tests to verify the updates.

Change-Id: I24196959a28093a292f599f93f6b78f6fd76c69e
Signed-off-by: Bernd Hufmann <bhufmann@gmail.com>
Reviewed-on: https://git.eclipse.org/r/10864
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>