tmf: Add a state system explorer view
This view displays raw state system information, synchronized with
the currently selected timestamp. This is amazingly useful for
debugging, but could also be useful for end-users trying to get
more low-level information.

For this reason, it's shipped with the runtime plugin, but not
present in the default perspective.

Based on initial implementation by Florian Wininger.

Change-Id: I2dc6b798862039b1c5c18ec5418d7f5913e27dcc
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/11628
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>