Work around the limitation in o.e.debug.ui.launchConfigurationTabGroups.
If 2 or more Launch Configuration Tab Group providers contribute to the
same 'type' and 'mode', they will clash, and the last one encountered
will be used, regardless of which launch delegate is specified. This can
cause problems when a launch delegate uses launch configuration
attributes that a user should be able to set with a particular tab.

We define 'org.eclipse.cdt.launch.applicationProfileLaunchTabGroup' as
the ID for the agreed upon tab group for all providers of this
functionality. (Bug 405207)

Change-Id: Idae4637a8bdb36015405f5762a7d06e3ccc86477
Reviewed-on: https://git.eclipse.org/r/11648
Tested-by: Hudson CI
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>