lttng: Correctly assign syscall state for new processes
If a process is already in kernel mode ("syscall" state) at
the beginning of a trace, all its children created via fork/clone
would be wrongly assigned to a user mode state once they get
scheduled on a CPU. This is because the parent's sycall name was
not set (since we never saw that syscall entry event).

This is actually a common case with LTTng traces: the lttng-consumerd
process is always already in kernel mode at the start of a trace,
and forks another process soon after.

Fixes bug #405221.

Change-Id: Iffdfb3277c1fa3fb9316c2ef9c8e6f04f539bb43
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/11756
Tested-by: Hudson CI