tmf: Insert null values for newly-created attributes
Instead of requiring every state provider to explicitely insert
a value when creating a new attribute, we can have the attribute
tree do it for them. Then, if the provider wants to set it to
something else, it's just changing a variable in memory.

This fixes a potential problem in the LTTng-kernel state provider,
where a value could get reset to null if no new attributes are
created between two sched_switch events.

Change-Id: I600125f560f0362671379848c572359be2f74cc7
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/12249
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>