tmf: Fix exception throwing in Sequence Diagram tests
Because of the new trace manager, the tests now have to send
"dummy" signals when traces are opened or selected, to register
them in the manager.

Also removed the RuntimeException when receiving a signal that
says to close a trace that is not currently open. We will simply
ignore it in this case.

Change-Id: I6ca345fb817a8a5691b41430b0062dbf5dc0812f
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/12441
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>