tmf: Add a Double type for state values
Double has a convenient Double.NaN value that cannot be attained
by "normal" means. We'll use this as the null representation of a
DoubleStateValue.

They are stored in the history tree the same way as Long state values:
by using a fixed 64-bit entry in the Strings section. If we start
using a lot of Long and Double values, eventually it might be worth
making the "valueOrOffset" field 64-bit to save space...

Change-Id: If59677565f16eaee947b9c4724fa7df7ef2239a9
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/16330
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>