tmf: Don't use .deleteOnExit() in unit tests
deleteOnExit() guarantees the file will be deleted *when the JVM exits*.
However, in most cases with unit tests, you want to clean up at the end
of the test, and not wait until the whole test run is over. This gives
more consistent behavior when running different tests in a different order.

So instead use a simple .delete() in @AfterClass methods.

Change-Id: I07f277dbe85b66659ca66c996b232cd64bb6868f
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/17136
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>