Tmf: Add the concept of module source to the Analysis API
This allows for other ways to define analysis modules than having to describe
them in the plugin.xml file.  For example, an XML file can be the source of an
analysis module.

The source itself needs to be described in the plugin.xml, using another
element of the analysis extension point.

Each module source will typically have its matching helper class to create the
instance of a module when required and to be able to access available modules
for different trace types before the trace is instantiated.

A default implementation of the IAnalysisModuleSource interface and its helper
are provided for the modules advertised through the extension point: the
TmfAnalysisModuleSourceConfigElement and TmfAnalysisModuleHelperConfigElement.

Change-Id: Ia2c7142704aa6b2a326bba9d2bf0734954427b7e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/19433
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>