tmf: batch import wizard smoke test
This is a sample of how to make a test using SWTBOT.
It should be a useful stepping stone to gain test
case coverage and control what we're reading.

The test features:
* a kernel trace generator
* importing a trace
* opening a trace
* navigation in a trace
* testing tracing perspective

What would be nice to have:
* closing a trace
* all the other functional test on all the trace types!

Please note that SWTBot needs to run on a non-UI thread.
https://wiki.eclipse.org/SWTBot/FAQ#Why_do_tests_run_on_a_non-UI_thread.3F
After the code was changed to run on a non-UI thread, it
was possible to remove many workarounds which made the code leaner and
easier to maintain. The UI code not handled by SWTBot still needs to run
in a UI thread, PlatformUI.getWorkbench() for example.

Change-Id: I91e258ed5fb7735287da303d518314460839547b
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/13780
Tested-by: Hudson CI