RPM: filter out %{?...} macros from outline
The eval function is calling the "rpm" command but to do this the plugin
tries to get a RemoteService to get the system environment variables.
The RemoteService is returning null because it cannot find a
remoteService extension point for RSE scheme.

For now it is temporarily better to filter out the %{?...} macros from
the outline entries rather than having NPEs and showing nothing at all.

Bug 421480

Change-Id: I941c2a33b4642097c87410ff860630a5f079a263
Reviewed-on: https://git.eclipse.org/r/18359
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>