tmf: Disable scheduler executionOrder test
That unit test has always been prone to multi-threaded uncertainty,
but recently it has been failing a lot more, which is very annoying
for the CI. I suspect it's because of the recent changes to the CTF
parser, some requests may end up finishing faster than they did before,
and before another request could be scheduled.

I suggest disabling it until we rewrite it to be more permissive (or
better, add testing mechanisms to the scheduler, and test the order
of requests being *sent*, not the order in which they finish).

Change-Id: I70db04feb9f86534f5d12623cf33f06d56b3a737
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/18942
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>